requirements.txt
playwright>=1.40.0
Flask>=3.0.0
pywebview>=4.4.0
APScheduler>=3.10.0
aiosqlite>=0.19.0
PyYAML>=6.0.1
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
